Equation and small code edits
I noticed that the variety C has the incorrect equation according to the picture drawn (and the paper published on this topic). Additionally, the function "differentiate" no longer takes a variable of type "System." This can either be changed by changing the function "differentiate(F, x)" to "jacobian(F, x)" or by making F into type Vector{Expression} as I have done. Note that "solve" still accepts this as a valid parameter so this simple fix should work.